refactor: change constructor function into a class to get a better typing
